requirements.txt
# Data Sources
akshare>=1.10.0
futu-api>=7.1.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.9.1
pylint>=2.17.5
mypy>=1.5.1

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.66.1

# Type Hints
typing-extensions>=4.7.1
